requirements.txt
streamlit
tensorflow
opencv-python
pillow
numpy
pandas
scikit-learn
matplotlib
icrawler
